Set controller ref to pod, not table
